requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.3.0
click==8.1.3
dj-database-url==2.0.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.11.2
importlib-metadata==6.7.0
isort==5.12.0
Markdown==3.4.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.0
zipp==3.15.0
